+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Streams
   * Hide FM Radio if hardware is not available
   * Make pandora stream album art use HTTPS urls to make sure it is rendered in the ios app
+  * Make Pandora streams a bit more robust against failure
 
 ## 0.3.3
 * Web App